Install BlockHound in integ-test-bases and move install test to service module
