+++ test/service-test-utils/src/main/java/software/amazon/awssdk/testutils/service/AwsIntegrationTestBase.java
@@ -17,6 +17,7 @@ package software.amazon.awssdk.testutils.service;
 
 import java.io.IOException;
 import java.io.InputStream;
+import reactor.blockhound.BlockHound;
 import software.amazon.awssdk.auth.credentials.AwsCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProviderChain;
@@ -27,6 +28,10 @@ import software.amazon.awssdk.utils.IoUtils;
 
 public abstract class AwsIntegrationTestBase {
 
+    static {
+        BlockHound.install();
+    }
+    
     /** Default Properties Credentials file path. */
     private static final String TEST_CREDENTIALS_PROFILE_NAME = "aws-test-account";
