+++ test/service-test-utils/src/main/java/software/amazon/awssdk/testutils/service/AwsTestBase.java
@@ -23,6 +23,7 @@ import java.io.InputStream;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
+import reactor.blockhound.BlockHound;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProviderChain;
 import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
@@ -30,6 +31,11 @@ import software.amazon.awssdk.core.exception.SdkServiceException;
 import software.amazon.awssdk.utils.IoUtils;
 
 public abstract class AwsTestBase {
+
+    static {
+        BlockHound.install();
+    }
+
     /** Default Properties Credentials file path. */
     private static final String TEST_CREDENTIALS_PROFILE_NAME = "aws-test-account";
